Modded actuator to cope with stepper signal generation. Modified the robot trajectory. Final commit during coupe de france 2025.

+++ 2025/robot/info/robot_core_src/match_mgr.h
@@ -44,7 +44,7 @@ void MatchMgrResetEventWaitforEndState();
 bool MatchMgrGetEventEndzoneState();
 void MatchMgrResetEventEndzoneState();
 
-void MatchMgrSetWaitingTimer(long waitingPeriodMs_u32);
+void MatchMgrSetWaitingTimer(uint32_t waitingPeriodMs_u32);
 void MatchMgrUpdateWaitingTimer();
 
 void MatchMgrChangeColor();
